Move test compiling bridges to independent suite
